pyproject.toml
[project]
name = "aiquerychat"
version = "0.1.0"
description = "LLM-guided SQL query tool with TUI chat interface"
requires-python = ">=3.12"
dependencies = [
    "click>=8.1.0",
    "sqlalchemy>=2.0",
    "pymssql",
    "any-llm-sdk>=1.0",
    "textual>=0.1.0",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
    "ollama>=0.6.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.9.0",
]

[tool.ruff]
line-length = 100

[tool.pytest.ini_options]
testpaths = ["tests"]
